Renamed 'gp_tools' and 'pdf_tools' modules to 'gp' and 'pdf'
