+++ docs/source/GpOptimiser.rst
@@ -2,14 +2,14 @@
 GpOptimiser
 ~~~~~~~~~~~
 
-.. autoclass:: inference.gp_tools.GpOptimiser
+.. autoclass:: inference.gp.GpOptimiser
    :members: propose_evaluation, add_evaluation
 
 Example code
 ^^^^^^^^^^^^
 
-Gaussian-process optimisation efficiently searches for the global maximum of a function by iteratively
-'learning' the structure of that function as new evaluations are made.
+Gaussian-process optimisation efficiently searches for the global maximum of a function
+by iteratively 'learning' the structure of that function as new evaluations are made.
 
 As an example, define a simple 1D function:
 
@@ -37,13 +37,13 @@ Create an instance of GpOptimiser:
 
 .. code-block:: python
 
-   from inference.gp_tools import GpOptimiser
+   from inference.gp import GpOptimiser
    GP = GpOptimiser(x,y,bounds=bounds)
 
-By using the ``propose_evaluation`` method, GpOptimiser will propose a new evaluation of the function. This proposed
-evaluation is generated by maximising an `acquisition function`, in this case the 'expected improvement' function.
-The new evaluation can be used to update the estimate by using the ``add_evaluation`` method, which leads to the
-following loop:
+By using the ``propose_evaluation`` method, GpOptimiser will propose a new evaluation of
+the function. This proposed evaluation is generated by maximising an `acquisition function`,
+in this case the 'expected improvement' function. The new evaluation can be used to update
+the estimate by using the ``add_evaluation`` method, which leads to the following loop:
 
 .. code-block:: python
